sprint 3

+++ POS/src/main/java/com/ust/pos/dto/CustomerDto.java
@@ -13,7 +13,7 @@ public class CustomerDto extends CommonDto {
     private Double balance;
     private String balanceType;
     private String partyType;
-    private Double creditLimit;
+    private Long creditLimit;
     private AddressDto billingAddress;
     private AddressDto shippingAddress;
 }
