+++ POS/src/main/java/com/ust/pos/model/Customer.java
@@ -15,5 +15,5 @@ public class Customer extends CommonFields {
     private Double balance;
     private String balanceType;
     private String partyType;
-    private Double creditLimit;
+    private Long creditLimit;
 }
